requirements/base.txt
Django==3.1.2
django-cors-headers==3.4.0
django-filter==2.3.0
djangorestframework==3.11.1
drf-yasg==1.17.1
Pillow==7.2.0
